A few minor details but important changed to get unit testing to work. Initially, on my Mac laptop, Jmult produced an obtuse error message in SP,   Error:  RMATxCVEC: matrix and vector sizes incompatible and SP2 forward solver failed with a segmentation fault with Intel compiler, unless run in debug mode.
1)  IMPORTANT: ioAscii.f90 changed to always use integer(4) for input and output of the electric fields in binary.
This may affect other users of the code who would have by now gotten used to the fact that this depends on the
system default. However, on a Mac laptop with an Intel compiler, it did different things on a whim (e.g., integer(4)
for cartesian code, integer(8) for spherical code in the gridelements version - not good for trying to use the files!)
Also, added some grid error checking on input.

2)  UserCtrl.f90 changed to add fwdCtrl input file to most sensitivity tests.

3)  Jmult segmentation fault in SP fixed by passing the interior edges only to Div in FWDsolve3D in EMsolve3D.f90.
Note that passing s(EDGEi) directly to Div (bypassing a temporary si complex vector) also causes the optimized version 
of the Intel compiled program to produce a segmentation fault. Not at all sure why, but it just took me hours to locate 
so won't do that again.

4)  COMPILER BUG FIX: ifort (IFORT) 19.0.4.233 20190416 produced a segmentation fault with SP2 only (not with SP!)
    at the first encounter of the line "B%val(j1:j2) = A(i)%val" in BlkDiag_Cmplx in module spOpTools.f90.
    Something to do with vectorization, since this doesn't occur in debug mode. Rewritten to use a loop
    with error checking, for now. This is most likely a temporary fix that could be eliminated later.
    Should not present a problem: this is only done once at the beginning. Note: error checking not necessary.
         !B%val(j1:j2) = A(i)%val
         do j = 1,size(A(i)%val)
            if (isnan(real(A(i)%val(j)))) then
                write(0,*) 'ERROR: NaN in BlkDiag_Cmplx for i=',i,' and j=',j
            endif
            B%val(j1+j-1) = A(i)%val(j)
         enddo

Now, Jmult in both SP and SP2 is still broken, but at least everything runs. Both versions pass all the other symmetry tests,
both in cartesian and spherical coordinates. SP2 runs much faster. Still haven't explored the accuracy and the rest.
The default matrix-free version passes all symmetry tests.

In all versions, is seems like the L operator is inaccurate in that the symmetry test is only passed if the numbers are large, 
and fails if the absolute value of the dot product is small. There's probably a subtle bug lurking somewhere. 

Everything only tested in serial, but I made no changes to the MPI.