A couple of bug fixes and tests that won't affect most "ordinary users".
1. Most importantly, the Jmult operation is *finally* working in SP(oid). Note that Jmult is currently only used in DCG (a.k.a. data-space CG) inversion. OK, most people probably do not know that ModEM supports DCG... Still, no excuses for this long-overdue fix -  
the problem turns out to be from a silly mistake that I divided the source term in the Jmult operation with an extra Vedge volume weight... 

the modified right hand side in the CCGD scheme *should* be: 
    s + {i\omega\mu}^-1 GD*s + b    
as in eq. [26] of the Dong and Egbert 2019 paper. the real discrete form is multiplied with the volume weight to (sort of) symmetrize the system: 
    - i\omega\mu V_E j - V_E GD_II j - V_E A_IB b
as we only know about the source term s(- i\omega\mu j), we need the j:
    j = s/(-i\omega\mu)
and multiply the volume weight after applying the modification term:
    V_E * GD_II * (s / (-i\omega\mu) 
simple enough - but I somehow thought that the s term was already multiplied with the Volume weight and had j = s/(-i\omega\mu)/V_E... 

the confusing variable names also contributed a little, so I spent some time to make the naming more consistent and sort out the program logic across the different SP versions (s, si, stemp, b). 

    NOTE to Naser: this should also fix the problem for any internal sources (i.e. sources from CSEM etc.). You can try the SP/SP2 with your CSEM testings.

    the Jmult operators are tested in SP and SP2 with GFortran. See the 3D_MT/SP*/EMsolve3D.f90 for details - I didn't have the time to try out the SPETSc2. it should work, though (I don't see why not).  

2. a new way to calculate the node-based conductivity is now implemented - 
   the new idea is to average the conductivity of the six edges (that connect to a certain node) instead of averaging the conducitivy of the 8 adjacent cells. This is more straight-forward and (somewhat) comprehensive when you are considering the divergence of the current from a node. thought this could bve useful for anisotropic cases.  
   see subroutine Cell2Node2 in 3D_MT/GridCalc.f90 for details

3. a pure cosmetic feature to keep track on the number of FWD/TRN evaluations for DCG inversion. the inversion scheme now outputs the total number of those evaluations when the iteration terminates. thought this could be useful when I was staring at the long output log of a DCG inversion.  
   see INV/DCG.f90 for details. 
   TODO: we may need a mechanism to stop the inversion when the DCG stalls. Now the inversion just goes on forever until the iteration number reaches the limit (or until you grow impatient and kill it).