This is an update (and explaination) to my previous quick fix
The main purpose is to (finaly) fix the CUDA support in the SPETSc2 forward module - need this to be compared with the "accelerated" solver that an Nvidia engineer and I have been fiddling with. 

For now the CUDA routine (fwdsolve3dc) gives (almost) identical results as the CPU-only ones - I have tested this on my workstation with a consumer GPU (GTX1080Ti), which is already slightly faster than my 8-core Xeon CPU in parallel, in terms of forward calculation.   

However, it is not very useful, in its current form, if you do not have a professional-grade GPU (i.e. full FP64 capibility). The consumer GPUs (like the one I have) are "nerfed" by design to limit the FP64 performance to 1/32 or 1/64 of their FP32 equivalents. In other words, if you have a professional card, you can expect a 16-32 times improvement in performance, in theory.  

Note that the PETSc lib is still very tricky to use, as ever - to utilize the CUDA, one key step is to configure PETSc with a cuda-aware MPI wrapper... Let me know if any of you are willing to test it!   

the detailed changlogs are listed below: 

MPI/Declaration.f90:
added a new logical variable of use_cuda and an integer variable of size_gpu to indicate the number of gpus in the whole system - useful if you have more than one gpus to spare 

MPI/Main_MPI.f90:
modified the worker_job subroutine to work with CUDA FWD/TRN methods, note that I have a really arbitary default layout for the grouping - the GPUs are now assigned to the last groups, which are supposed to deal with longer periods. Note this is not yet Benjamin-safe, as the periods are not necesarily sorted in ascending order. 

3D_MT/ForwardSolver.f90: 
slightly modified to work with the now stand-alone CUDA routine 

3D_MT/FWD_SPETSc2/EMsolve3D.f90: 
fixed a couple of variables that cause the illegal memory access in the PETSc routine (FWDsolve3Dp) 
Also introduced a new standalone CUDA procedure (fwdsolve3dc) to the fwdsolve3d interface. The cuda support was already in the FWDsolve3Dp procedure for more than a year ago, but never tested properly. 
To tell the truth, I don't think it is wise to make the cuda precedure a stand-alone one. This is however easier to debug and test (without breaking the other functions) - I will keep it here until I get an idea on the management of the (increasingly) heterogeneous computation infrastructures 

3D_MT/FWD/EMsolve3D.f90: 
3D_MT/FWD_SP/EMsolve3D.f90: 
3D_MT/FWD_SP2/EMsolve3D.f90: 
All modified to add "use_cuda" dummy variable to be compatible with the new MPI routines. 

CONFIG/Configure.SPETSc2.MAC.Intel
added a new configure file sample for SPETSc2 - one can try
    ./CONFIG/Configure.SPETSc2.MAC.Intel Makefile MPI
to prepare CPU-only makefiles, or use  
    ./CONFIG/Configure.SPETSc2.MAC.Intel Makefile CUDA 
to prepare a template for cuda (one still needs to manually supply the PETSC lib information). 

TODO: working on a more reasonable stratagy for heterogeneous system with multi-cpu and multi-gpus. Also - maybe I *should* get a better GPU, after I try out the A100 GPU that NVidia provides...