Some long overdue update to the D-ILU preconditioner, now one can finally direct-compare the result from MF/SP and SP2 - (before this, the SP2 has to use my homebrew ILU0 preconditioner, as the old SP version of D-ILU routine does not support assymetric matrices).
This also address (some of) Gary's concerns about the preconditioning matrices for different Txs. In this case, only the diagonal part of the U matrice need to be modified for a different freqeuncy/sigma.  
So one can write a clever routine to update the matrices - instead of assembling them again. But I should say that one still needs to scale the lower triangle matrix L with the new diagonal. This probably will not save you *too much* time... 

TODO: check if the sparsity pattern will change with a new sigma in CCGD. 

3D_MT/SP_topology/spOpTools.f90
fixed the dilu_complx_AS routine to work with asymmetric matrices (i.e. the system equation of CCGD in SP2/SPETSc2). I guess I wrote the AS version a few years ago, but never made it working... 

3D_MT/FWD_SP2/modelOperator.f90: 
change the default preconditioner to D-ILU_AS to be consistent with the SP and MF versions.  

3D_MT/FWD_SPETSc2/modelOperator.f90: 
change the default preconditioner to D-ILU_AS to be consistent with the SP and MF versions.  
note this does not actually change the performance of the CUDA solver, as they do not have a D-ILU support in CUDA, yet!

3D_MT/FWD_SPETSc2/EMsolve.f90: 
some minor modifications for the CUDA solver - the i/omega/mu/sigma diagonal is now manipulated in PETSc routines (instead of SP routines)