A few updates for the solver/MPI modules prior to the major update for GPU solvers, people who are using CPU-only calculations can (hopefully) cling to this version for some time, in case my GPU update leads to some turmoil in other part of the code.
3D_MT/FWD/solver.f90 
3D_MT/FWD_SP/solver.f90 
3D_MT/FWD_SP2/solver.f90 
3D_MT/FWD_SPETSc2/solver.f90 

added a long overdue "working" TFQMR solver for all SP methods - 
this is mainly for the SP2/SPETSc2 version as the system matrix in CCGD is
no longer symmetric - the old QMR method will need the transpose of CCGD,
which is not saved in current set-up to reduce the memory consumption. 
This leads to two extra transpose operations in a QMR iteration, which slows
down the calculation.  
also omits an unnecessary SPMV operation in the BiCG code (for all FWD
methods). Should save some time for BiCG iterations, also. 

3D_MT/FWD/EMsolve3D.f90 
3D_MT/FWD_SP/EMsolve3D.f90 
3D_MT/FWD_SP2/EMsolve3D.f90 
3D_MT/FWD_SPETSc2/EMsolve3D.f90 

updated the interfaces for various fwdsolve3d subroutines to include the
device_id variable to allow for the CUDA device manipulation

3D_MT/ForwardSolver.f90
2D_MT/ForwardSolver.f90

updated the interfaces for fwdsolve and sensolve subroutines to include the
device_id variable to allow for the CUDA device manipulation
note the 2D_MT part is merely for compatibility, there is no GPU support for
2D yet (it is useful to introduce GPU based direct solvers, though)

MPI/Main_MPI.f90
MPI/Declarition_MPI.f90

added a couple of methods to determine the number of devices on local nodes -
and a few fixes regarding using multiple CPUs to feed the tasks for a single
GPU. It also works for multiple GPUs on a node, as well.  

TODO: add a method to work with multiple GPUs one multiple nodes (need to 
find the money to buy some, though)

3D_MT/Sub_MPI.f90
2D_MT/Sub_MPI.f90

corrected a typo of "count_number_of_meaasges_to_recv", well...