A couple of important updates on the two-layered parallel structure (as discussed in the GPU paper).
MPI/Declaration_MPI.f90, Main_MPI.f90: 
The most important feature is a topology-based parallel paradigm. We can now get to know the number of GPU devices on a physical node, which allows an easier multiple GPU usage on multiple machines. 
The other feature is a node-based parallel method, controlled by a master switch (para_method) in Declaration_MPI.f90. The basic idea is to use the bandwidth of one entire server node to, well, accelerate one single FWD/ADJ task. It allows one to utilize tens of computational servers at once (with the PETSc routines, of course). That’s probably the best way to get maximum throughput if one doesn’t have any GPUs at hand. 
That can, of course, singlehandedly dry out your CPU-hour deposit - but I guess it worth it if you can get a total bandwidth with tens of Terabytes and get the job done faster. But note that this is turned off on default (set para_method = 1 and re-compile to try it). 

3D_MT/FWD_SPETSc2/ EMsolve3D.f90, modelOperator3D.f90: 
I also tried to make the PETSc version compatible with the new configuration format. Not sure if it works correctly, though. 

Makefiles: 
updated the tempelate makefiles (probably only used by myself for quick switch of branches).