3D_MT/SP_Topology/spOpTools.f90 3D_MT/SP_Topology/vecTranslate.f90 3D_MT/FWD_SP2/EMsolve3D.f90 3D_MT/FWD_SP2/modelOperator3D.f90 3D_MT/FWD_SP2/solver.f90
A fix for a long-standing problem regarding the SP2 segmentation fault
errors - or maybe I should say an attempt to fix, as I am not sure if 
it really works yet - it seems to work on my side, though. 

I guess the problem dates back to the days when Benjamin was still in OSU. 
This is almost a myth now as there are various reports about how 
the segmentation problem would emerge, on different situations. I wasn't 
even sure this is a bug, as I was never able to replicate his problem. Ben 
and Anna's test cases seemed to (somehow) run just fine on my machine.

Now that Anna brings it up again with the spherical coordinate, I spent some 
time to dig into the issue. As I never encountered that problem with 
Gfortran, my intuition is that the problem has something to do with the 
Intel compiler, which both Anna and Ben use. 

With the help of Valgrind I found that we can get an out of boundary error, 
related to the G -> T transpose when deciding the air and earth edge/nodes,
but only with intel fortran.  

Note this (the air/earth term seperation) is only used in the CCGD (SP2) 
version, for the node-based regularization term. So there is a possibility 
that the earlier versions based on edge-based regularization are not affected 
by this - I have not tested this, though. One may try to test it, with the
GradDivSetup1 subroutine. Simply replace the default GradDivSetup with 
GradDivSetup1 - and see what happens!   

Anyway, it turns out that we can trace the problem back to the original 
gridIndex subroutine in the vecTranslate module. That subroutine does a job 
to convert the i,j,k indices (from the original mf version) from an 1D index 
to used in sparse matrices. 

And that's where we have a problem with intel fortran, as Gary designed the 
subroutine to use float variables to determine the indices. For example, if we have Nx by Ny by Nz nodes - and the 1d index is ii

we can use mod(ceil(float(ii)/float(Nx)),Ny) to find the Y indices.

normally this should be totally fine - but the round-up error seems to 
be a prolem when we have very large meshes, and when the intel fortran 
tries to use vectorization to calculate the index. 

so, we tend to get the "out of bound" problem, when we have a value like 
38.0000...002, which will be interpreted as 39, with the ceil function.

to deal with this I resort to the integer calculations...
using  mod((ii+Nx-1)/Nx,Ny) instead...

So far the fix seems to work fine, with both Intel and Gfortran, but I am not 
sure if there are other places like this - which we overlooked for years.  

Anna - please let me know if this solves your problem, thanks!

TODO: also need to look for memory leaks - don't know if they lurk 
somewhere...