3D_MT/ForwardSolver.f90 2D_MT/ForwardSolver.f90 MPI/Main_MPI.f90
A fix for a strange problem regarding the parallel TRN calculation in 
inversions. This affects all versions with MPI calculation (even for 
2D, I assume). I found that problem when I was testing my new 8-card GPU 
platform. There is a strange lag when the master distributes tasks to 
all the workers and send the corresponding background fields in a loop.
I believe this is a problem we overlooked for years.  

It turns out the initSolver operation excuted by each worker takes too
long (mainly because of the decomposition for the preconditioner) - 
so that the MPI_SEND transfer of the field e0 cannot start until the 
initSolver finishes. As this happens, the loop for each worker will get
stuck at each initSolver call. This can easily leads to a lag of >
100s if you have a large system, and many transmitters/periods

the solution is also simple - just transfer the e0 before the initSolver
routine. In that case the master will just turn to the next worker and 
leave the previous worker doing the initialization. no further lags in 
this case. To do this I had to add a new "initSolverWithoutE0" routine
(for both the 3D and 2D!) to do the preparison without tampering the e0 
transfered from the master.


I didn't realized this as previously the TRN calculation on CPUs are 
slow enough so that the lags were not that pronouced. However, for now
the TRN is solved in <10s with GPU. This can lead to a inferior 
performance as the lag can take even longer time than the actual solver!


TODO: also need to look through the FWD works