3D_MT/FWD_SP2/solver.f90 3D_MT/FWD_SP2/hipFortMap.f90 3D_MT/FWD_SP2/kernel_c.hip MPI/Main_MPI.f90 MPI/Declaration_MPI.f90
A major update to support the non-Nvidia GPUs through HIP interface,
which is developed mainly by AMD (and a bunch of "fan" workers). 
HIP is essentially a mimic of the CUDA interface that utilizes almost
identical (can be quite different in some ways) grammar and keywords.
The idea is use HIP as a bridge to use AMD and Intel (plus some 
Chinese brands') GPUs, not everyone has a Nvidia GPU, anyway.

Still, porting from CUDA to HIP can lead to potentially strange 
behaviors, as these are indeed built on different architectures. 
I had to add (yet) another two files: the "kernel_c.hip" for custom
kernels and the "hipFortMap.f90" for c-fortran interfaces for HIP.

tested the program with GCC+HIP 6.1.0 on ubuntu 20.04. So far it is 
getting identical results and a decent speed-up with AMD and Chinese 
Hygon GPUs (not as great as Nvidia cards, though). Please see the 
Makefile.hip and Makefile.cuda for examples to setup the compiling 
environment.

Although I haven't changed too much on the Nvidia side, this update
could still break the already working CUDA parts (I made some changes
in the CUDA part to get consistent behavior between the two). 

Please use this with caution, and roll back to the previous one, if 
needed. 

Let me know if you encounter anything bizarre with CUDA or HIP.

Hao

TODO: guess I need to design a uniform driver bicg solver for both cuda 
and hip