3D_MT/FWD_SP2/EMsolve3D.f90 3D_MT/FWD_SP2/modelOperator3D.f90 3D_MT/FWD_SP2/solver.f90 MPI/Main_MPI.f90 MPI/Declaration_MPI.f90
A major update to prepare the code to merge with my multi-GPU branch.
Essentially this is a new fine-grain parallel framework, currently for 
CPUs only. It uses the similar two-layered parallelization conception 
I used in the SPETSc2 branch - albeit without any PETSc libaries 
(it's completely native in Fortran now). 

Here we split the system and preconditioner matrices into row blocks and 
diagonal blocks, before letting different CPU processes to deal with each sub-
matrix - the corresponding MPI scheme is also changed, with the same idea 
used in the multi-GPU branch.

I implemented this for fine-grained parallel methods for both CPU and GPU. 
I guess at some stage we may dump the PETSc version, which is already kind 
of obsolete - but it may still be helpful if people wants to test some 
interesting new solvers/preconditioners without writing their own routines

Currently it should be useful in practical for people who have a lot of CPU
clusters at disposal, but don't have decent GPUs... I will use this for 
the direct comparison between fine-grained CPU and GPU routines. 

For those who wants to test - please use the -DFG macro to enable this with
SP2 version. See Makefile.sp2fg for a configuration example.

The code has been tested with open-mpi 4.1.2 and gcc 9/11 on my 16-node
cluster (and my MBP). 
So far the best cost-effective scheme is still using 3 procs
per period/polarization (roughly the same as I figured out with the NCAR
system, for the old SPETSc2 version). This allows you to use nTx * 3 
cpu processes for each FWD/ADJ task, or 150 CPU cores for 25 periods and 
2 polarizations.

Please note that you really need to spread your processes out to more nodes
to get good speed-ups, e.g: distribute 6 processes on each of 10 nodes, 
instead of 30 processes on two nodes each. 

Or, as I pointed out in the original GPU paper, more processes on a single
node would simply saturate the memory bandwidth, with little efficiency 
boost. In other words, do not use the FG version if you don't have a lot of 
cluster nodes. 

I would not give more than 6 procs to a Tx task, as the block diagonal
preconditioner detoriorates very quickly with the number of sub-blocks. 

TODO: one possible idea, to improve the efficiency of the block ILU
decomposation scheme, is to distribute the "local" decomposation tasks 
to workers -
However, it may break a thing or two to implement this - so I kept it 
simple to let the group leader do all the works (as in the previous version)