Update UpdateFreq subroutine interface to match new SP2 UpdateFreq
In a pervious revision, a change was made to the subroutine `UpdateFreq` in the
`3D_MT/FWD_SP2/modelOperator3D.f90` file. This change added a new argument to
`UpdateFreq`, `nproc`. However, this change was not mirrored in the `UpdateFreq`
subroutine of the `3D_MT/FWD/modelOperator3D.f90` version of the code.

This caused a compilation error when building the `FWD` directory, as calls to
`updateFreq` had too many arguments.

Thus, this commit adds that new, optional argument to the `UpdateFreq`
subroutine in `FWD/modelOperator3D.f90` to avoid that compilation issue. The
new argument performs no operation, and, because it is unused, will cause an
unused dummy argument warning to appear if compiling with relevant compiler
warning flags.

There is probably a better way to resolve this issue. Mainly, the `UpdateFreq`
in `FWD_SP2` could, perhaps, call MPI (or an MPI wrapper library) to 
inquire about the number of 
tasks (e.g. `call MPI_Comm_size(MPI_COMM_WORLD, ntasks, ierr)`.