Add ModEM_timers module to provide higher resolution timers to ModEM
This commit adds a new Module to ModEM that provides high resolution, named timers to ModEM
developers.

Previously, ModEM used the `date_and_time` intrinsic for timing operations. This works
for basic timing; however, the clock provided by `date_and_time` is not monotonic and can
speed up and slow down, mostly from NTP (for more details see:
https://blog.habets.se/2010/09/gettimeofday-should-never-be-used-to-measure-time.html).

Thus, this new module uses monotonic clocks. By default, it will use Fortran's
`system_clock` routine, but has the ability to use C timers. In order to use the C timers
however one will need to use the `-DUSE_C_TIMERS` compiler flag, and follow the directions
in `ModEM_machine_timers.c` to add it to your Makefile.

With the monotonic clock, the ModEM_timers module easily allows the ability to create,
start, stop, and destroy a new timer by a given name. It also has routines to generate a
report of all timers, or to print out the value of a timer at a specific time.

This will provide ModEM developers a way to easily add high resolution timers to sections
of the code, without having to worry about the details of timer types or low level timers.

Lastly, this commit adds new timers in a few places in ModEM. In this commit, that is:

* Total Time - Total time of running modem
* Forward - Total time of forward operation
* Inversion -  Total time of inversion
* NLCG Iteration - Time of each NLCG iteration