Actually add in ModEM_timers files
Mistakenly, revision 779 did not actually add the files of the ModEM_timers,
thus, this commit adds in those files.

As well, this commit destroys the 'NLCG Iteration' timer at the end of the NLCG
inverse operation so that the 'NLCG Iteration' timer is not included in the
final timer report.