ModEM_Timers - Add in unit testing and C timer information
This commit adds in a file to be used for Unit testing the ModEM_timers module.
To use it, you will need to compile the ModEM_timers.f90 and
ModEM_machine_timers.c files seperatly from ModEM and link with
ModEM_timing_tests.f90. The unit tests should be ran whenever changes to the
ModEM_timers module are made.

This commit also includes information on how to compile and use the C timers
with ModEM_timers.f90