Update errStop to call MPI_Abort rather than just stop
This revision updates errStop to call a new subroutine, ModEM_abort. ModEM_abort
is similar to the previous version of errStop in that it stops the model, but
ModEM_abort calls MPI_Abort and calls stop with a non-zero argument.

Previous to this commit, errStop would just call `stop` without any stop codes.
Calling `stop` alone causes 'normal' termination of a program and returns an
exit code of zero.

This causes problems when running the model with MPI and with a job scheduler,
such as PBS or SLURM, as the job scheduler will not cancel the HPC allocation
after a call to `errStop`. Such an event can waste valuable core hours as the
model could fail early on.

So far, I have mostly seen this happen when the incorrect amount of arguments,
or an incorrect job flag (e.g. -f rather than -F), but there might be other
places where such an event occurs.

Thus, this revision updates `errStop` to call `ModEM_abort()` to then call
`stop` with a non-zero argument to signify success or failure to the calling
program.  Likewise, if compiled and running with MPI, `ModEM_abort()` calls
`MPI_Abort(MPI_COMM_WORLD, ...)` to ensure that all MPI tasks will be
terminated. While this will prevent a batch job from exiting quickly on error,
it will also help prevent hanging MPI tasks on non-HPC systems.

Finally, this revision replaces calls to `stop` UserCtrl.f90 with either
`errStop` or `ModEM_abort()`. There are other places in ModEM where stop is
being used. A future commit should replace such calls with either a call to
`errStop` on failure or to a subroutine, which is yet to be created, that
stops the model successfully.

Properly returning a non-zero exit code on failure will also help with future
work for continuous integration.