Remove MPI communicator check and reassignment that created inverse deadlock
This commit removes the check and assignment of MPI communicators in
Master_job_Distribute_Taskes and Master_Job_Regroup that was causing a deadlock
condition when running the inverse with MPICH.

This was because the check was assuming that a negative value for the
communicator would mean an erroneous state; however, it seems that MPICH is
fine having a negatived valued communicator.

Thus this commit removes that check and reassignment. In other commits, we
should move towards using MPI types instead of integers for defining our
communicators. This might require we add some additional types to obfuscate the
communicator to handle running without MPI.