Remove uncommented header of fmkmf.pl and create default Makefiles
This commit removes the commented 'header' code that was generated by fmkmf.pl.
While these options may be helpful to some users, it makes the created
makefiles messy and hard to read. Users should add their desired flags as they
see fit.

This commit also adds `iso_fortran_env` to fmkmf.pl so that a warning is not
displayed that the file cannot be found.

Lastly, this commit removes example makefiles and replaces them with two
example makefiles, both for the matrix-free version and one for GNU and one for
Intel.

We should think about how we manage or communicate how to compile the SP, SP2,
SP2 `-DFG`, `-DHIP`, `-DCUDA` versions of ModEM.

I'm hesitant to include a list of Makefiles, as they easily get out of date and
it makes code changes hard, as the developer needs to create a whole host of
Makefiles.

Perhaps it would be easier to update or add configuration files to take in an
argument which will allow one to compiler with `-DFG`, `-DHIP` or `-DCUDA`?

with our users, mainly, SP, SP2, FG, HIP and CUDA.