Do not use FWDSolve3Dp if FG is not defined
The SP2 forward solver had a compilation bug that would not allow it to compile
without -DMPI. This is because the subroutine FWDSolve3Dp was being compiled,
which contains MPI variables and functions.

This hot fix only defines that subroutine (and calls to that subroutine) if
-DFG is defined. We might need to do additional testing and only allow this
subroutine to be built if MPI and FG are built, but for now, this should
resolve Anna's issue.