Fix a major inefficiency that resulted in insufficient VM error
fix a major inefficiency that resulted in weird insufficient virtual memory
errors on reading very large data files; the fix involves pre-allocating the
receiver dictionary for 200,000 receivers, then truncating after the data are
read. If you need more receivers, it should catch that and exit with a warning
to change that maximum in the code. Should be good enough for most purposes.