Allow SP2 to run in serial (without MPI)
This commit allows the SP2 forward implementation to run in serial without MPI.
A previous attempt to fix, which is not in the main or develop branch,  caused
a segmentation fault when running with a -DFG version with nTasks < nTx as that
fix would cause it to use the FG FWDSolve3D even if the FG code fell back onto
the non-FG code.

This fix does not have that issue and allows the FG code to either select
FWDSolve3Dp or FWDSolve3Ds (via the FWDSolve3D interface). It should allow for
normal operation otherwise. However, there might be a better fix and this might
just be a temporary solution.