Replace all usages of 'stop' with 'ModEM_Abort' or 'errStop'
This large commit replaces all* Fortran stop / STOP statements to either
`ModEM_Abort` or `errStop` defined that are defined in utilities.f90.

In conjunction with commit 7a29f11, this will ensure that `MPI_Abort` is called
when an error occurs. This will ensure that all tasks are properly stopped when
an error occurs. Before this, MPI tasks would often continue to run after an
error occurred until they were killed by an outside signal.

This will prevent these zombie tasks from taking up memory when running on
local machines and will prevent wasted core hours if running on a machine with
a job scheduler. With 7a29f11, modem will send a non-zero exit code, which will
alert the job scheduler that an error occured and will cause the job scheduler
to kill the job early, preventing possibly wasted core hours.

This change will also help in other applications which might call ModEM
programmatically, such as testing code. Calls to kill the model in the event of
an error will now return an non-zero exit code that can alert the calling
program of failure.