(r790) Another major update to merge the multi-GPU branch to stable-
Updated the cudaFortMap and kernel_c part to incorperate the NCCL library for
multi- card communications. Note this only works for NVIDIA cards, for now, as
I have not yet included the RCCL from HIP libs (which is a mess, as the HIP
code is quite several versions behinde the CUDA version. I need to collect
myself to find courage to merge the HIP branch... Probably won't be doing that
before I finish the multi-gpu manuscript...

To make the four types of solvers (CG,CG+GPU,FG,FG+GPU) compatible, I also
revised the interface of FWDsolve3D (again) - now we have a "FWDsolve3D"
interface for coarse grain CPU and GPU solvers, and a "FWDsolve3Dfg" interface
for fine-grain CPU and GPU solvers. Have not yet tested if this breaks the
serial code.

Please refrain to update to this version for production - I have not yet
fully tested this - might break a few things (e.g. the normal fg). I have
seen instances of lock or freezing of ncclComm apis on some machines - not
sure if that's a problem of the lib or the code.