(r793) A quick (and dirty) fix for a bug that prevents the (coarse-grain) parallel
GPU versions to use more than one GPUs (which is probably introduced after
r774).  The problem came from the old scheme to count the number of CPUs that
can utilize GPUs. In an ideal circumstance, the processes from a physical node
should count the number of CPUs and GPUs on that node, before determining which
processes should get a GPU. This is kind of hard to calculate, without
introducing extra data structure - currently I just use a dirdy way to fill the
node topology data in the rank_local and size_local variables - which are not
really used in coarse-grain methods. Of course this is not very elegant...

TODO: think of a better way to do the topology determination.