feature: support multi-node heterogeneous parallel
introduced 'size_node' and 'rank_node' parameters to let the procs know about the location they are in a physical node and try to determine the affinity of GPU devices in that certain physical machine. This is useful when you have some really heterogeneous layout (i.e. each node has different number of devices, etc - this 'number' can be zero)  note that we still do not have any means to get to know about the NUMAs (and the relavent CPU affinity)