Add a Dockerfile that is able to compile and run ModEM
This commit introduces a Dockerfile to the ModEM repository in an Ubuntu
container. So far, the Dockerfile installs LAPACK, LABLAS, GCC-14, GFortran-14,
g++-14 and 'manually' builds MPICH-4.3.1 along with a few other helpful
programs to have around.

I needed to configure and compile 'manually' here, rather than use `apt-get` as
the `apt-get` binary is configured with the `--with-pm=pmix`. Apparently, newer
versions of Ubuntu come with and updated PMIX which can cause issues when
running mpiexec, thus, we needed to configure MPICH to use hydra here instead
of PMIX.

It should be pretty easy to add additional libraries here if we desire (CUDA,
HIP, HDF5 etc.), but this will be a good starting point.

For reference, I've mostly done created this so I can use GDB in a non-slurm
environment. Apple silicone often doesn't produce meaningful traceback messages
for Fortran and I am tired of fighting with slurm, srun and Cray systems
shenanigans.