Convert User Guide to sphinx documentation
This commit converts the ModEM User's Guide to a Sphinx documentation. It can
easily be built if you have sphinx installed:

```bash
$ pip install sphinx
```

Once you have it installed, you can call `make html` from the `doc/` directory
and the docs will be built!

``` bash
$ cd doc/
$ make html
```

We still don't have a site to host this yet, but a future commit/work will
implement that.

I think the plan now is to use github pages. It should be a little easier to
maintain then, but will require a bit of overhead to get a github action to do
the heavy lifting, but it wont be too hard.