Convert Users Guide to Sphinx documentation and add automated github build/deploy action for it (#28)
This Merge converts the ModEM PDF User's Guide from a TEX/PDF to sphinx. So far, most of the User's Guide has just been copied over with a few added sections.

I took some liberty and spread out the 'Change Log for Release 1.2.0, April 12, 2019' section into a few different sections so they are easily findable.

There are also a host of new sections that were not previously in the User's Guide.

This merge also adds in a GitHub action that will automatically build (and deploy to modem.github.io if it is on main) this sphinx documentation. If it's run on main, the workflow will checkout main, build it, and then push the built files to the gh-pages branch, automatically updating modem.github.io!