Add in a single Configuration file which can generate most ModEM Makefiles (#25)
This merge adds in a single configuration file, Configure, which has the ability to generate make files for all solvers, compilers and ModEM build options.

There is already a lot of description in the Configure file itself, so no need to re-hash it here, but the goal of this single configure file is to make our lives easier as developers. Rather than making changes to all 30 something configuration files, we now only need to manage one.

This Configure file only supports GFortran and Intel Fortran compilers, but could easily be added to support more.

Currently, this configuration file does not have the capability to generate HIP/Cuda compatible makefiles. Perhaps that enough is a reason to use a different build system, currently, we are quite limited with building anything that is not a Fortran module.

Regardless, this should be easier to maintain, and easier for users to understand.

This script does make some assumptions, by default it will use MPI, rather than serial, and will build SP2, if the solver method is not specified. In the future we might want to make more assumptions, or could test to see if the user has specific compilers avaliable, i.e. check for gfortran, ifort, mpifort, but I don't think we have the number of new-users to justify that. We can always make that change if we get feedback that ModEM is hard to configure.